Fix branch resolving behavior, extract less relvant code from BlockImporter
