+++ src/main/scala/io/iohk/ethereum/utils/ByteStringUtils.scala
@@ -1,0 +1,8 @@
+package io.iohk.ethereum.utils
+
+import akka.util.ByteString
+import org.bouncycastle.util.encoders.Hex
+
+object ByteStringUtils {
+  def hash2string(hash: ByteString): String = Hex.toHexString(hash.toArray[Byte])
+}
